Refactor ServiceRequestsRelationManager and ViewServiceRequest to utilize ServiceRequestInfolist for improved schema configuration
