+++ app-modules/service-management/src/Filament/Resources/ServiceRequests/Pages/ViewServiceRequest.php
@@ -36,12 +36,7 @@
 
 namespace AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages;
 
-use AidingApp\Contact\Filament\Resources\ContactResource;
-use AidingApp\Contact\Models\Contact;
-use AidingApp\Division\Models\Division;
-use AidingApp\ServiceManagement\Actions\ResolveUploadsMediaCollectionForServiceRequest;
 use AidingApp\ServiceManagement\Enums\ServiceRequestTab;
-use AidingApp\ServiceManagement\Enums\SlaComplianceStatus;
 use AidingApp\ServiceManagement\Filament\Actions\ReclassifyServiceRequestAction;
 use AidingApp\ServiceManagement\Filament\Concerns\ServiceRequestLocked;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\RelationManagers\AssignedToRelationManager;
@@ -49,8 +44,8 @@ use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\RelationManag
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\RelationManagers\CreatedByRelationManager;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\RelationManagers\ServiceRequestConversationsRelationManager;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\RelationManagers\ServiceRequestUpdatesRelationManager;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Schemas\ServiceRequestInfolist;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\ServiceRequestResource;
-use AidingApp\ServiceManagement\Filament\Widgets\ServiceRequestMediaTable;
 use AidingApp\ServiceManagement\Models\ServiceRequest;
 use AidingApp\ServiceManagement\Models\ServiceRequestAssignment;
 use AidingApp\ServiceManagement\Models\ServiceRequestFeedback;
@@ -60,21 +55,15 @@ use AidingApp\Timeline\Livewire\TimelineList;
 use App\Enums\Feature;
 use App\Filament\Concerns\FiltersManagersFromGroups;
 use App\Settings\DisplaySettings;
-use Carbon\Carbon;
-use Carbon\CarbonInterval;
 use Filament\Actions\EditAction;
 use Filament\Infolists\Components\TextEntry;
-use Filament\Infolists\Components\ViewEntry;
 use Filament\Resources\Pages\ViewRecord;
-use Filament\Schemas\Components\Grid;
-use Filament\Schemas\Components\Group;
 use Filament\Schemas\Components\Livewire;
 use Filament\Schemas\Components\Section;
 use Filament\Schemas\Components\Tabs;
 use Filament\Schemas\Components\Tabs\Tab;
 use Filament\Schemas\Schema;
 use Illuminate\Support\Facades\Gate;
-use Illuminate\Support\HtmlString;
 use Livewire\Attributes\Url;
 
 class ViewServiceRequest extends ViewRecord
@@ -98,15 +87,15 @@ class ViewServiceRequest extends ViewRecord
     {
         return $schema
             ->schema([
-                $this->headerSection(),
+                ServiceRequestInfolist::headerSection(),
                 Tabs::make()
                     ->columnSpanFull()
                     ->livewireProperty('tab')
                     ->tabs([
                         ServiceRequestTab::Request->value => Tab::make(ServiceRequestTab::Request->getLabel())
-                            ->schema($this->requestTabSchema()),
+                            ->schema(ServiceRequestInfolist::detailSections()),
                         ServiceRequestTab::Files->value => Tab::make(ServiceRequestTab::Files->getLabel())
-                            ->schema($this->filesTabSchema()),
+                            ->schema([ServiceRequestInfolist::filesSection()]),
                         ServiceRequestTab::Assignments->value => Tab::make(ServiceRequestTab::Assignments->getLabel())
                             ->visible(fn (ServiceRequest $record): bool => filled($this->assignmentRelationManagers($record)))
                             ->schema(fn (ServiceRequest $record): array => $this->assignmentRelationManagers($record)),
@@ -157,176 +146,8 @@ class ViewServiceRequest extends ViewRecord
         ];
     }
 
-    private function headerSection(): Section
-    {
-        return Section::make()
-            ->heading(fn (ServiceRequest $record): HtmlString => new HtmlString(
-                view('filament.infolists.components.service-request-heading', [
-                    'serviceRequestNumber' => $record->service_request_number,
-                    'category' => $record->category,
-                    'type' => $record->priority?->type()->first()?->name,
-                ])->render()
-            ))
-            ->schema([
-                TextEntry::make('division.name')
-                    ->visible(fn (): bool => Division::count() > 1)
-                    ->label('Division'),
-                Grid::make(3)
-                    ->schema([
-                        TextEntry::make('respondent')
-                            ->label('Customer Contact')
-                            ->color('primary')
-                            ->html()
-                            ->state(function (ServiceRequest $record): string {
-                                /** @var Contact $respondent */
-                                $respondent = $record->respondent;
-                                $organizationName = $respondent->organization->name ?? 'Unaffiliated';
 
-                                return "{$respondent->{Contact::displayNameKey()}} ({$respondent->type->name})<br>{$organizationName}";
-                            })
-                            ->url(function (ServiceRequest $record) {
-                                /** @var Contact $respondent */
-                                $respondent = $record->respondent;
 
-                                return ContactResource::getUrl('view', ['record' => $respondent->id]);
-                            }),
-                        TextEntry::make('created_at')
-                            ->label('Created')
-                            ->dateTime()
-                            ->hintIcon(null),
-                        TextEntry::make('updated_at')
-                            ->label('Last Updated')
-                            ->dateTime()
-                            ->hintIcon(null),
-                    ])->columns(3),
-                Grid::make(3)
-                    ->schema([
-                        TextEntry::make('status.name')
-                            ->label('Status')
-                            ->badge()
-                            ->color(fn (ServiceRequest $record): string => $record->status->color->value),
-                        TextEntry::make('priority.name')
-                            ->label('Priority'),
-                    ])->columns(3),
-            ])
-            ->columns();
-    }
-
-    /**
-     * @return array<Section>
-     */
-    private function requestTabSchema(): array
-    {
-        $formatSecondsAsInterval = fn (?int $state): ?string => $state ? CarbonInterval::seconds($state)->cascade()->forHumans(short: true) : null;
-
-        return [
-            Section::make('Title')
-                ->schema([
-                    TextEntry::make('title')
-                        ->hiddenLabel(),
-                ]),
-            Section::make('Description')
-                ->schema([
-                    TextEntry::make('close_details')
-                        ->hiddenLabel()
-                        ->markdown(),
-                ]),
-            Section::make('Form Details')
-                ->visible(fn (ServiceRequest $record): bool => ! is_null($record->serviceRequestFormSubmission))
-                ->schema([
-                    Grid::make(3)
-                        ->schema([
-                            TextEntry::make('serviceRequestFormSubmission.submitted_at')
-                                ->label('Submitted')
-                                ->dateTime(),
-                            TextEntry::make('serviceRequestFormSubmission.author.email')
-                                ->label('Submitted By')
-                                ->color('primary')
-                                ->url(function (ServiceRequest $record): ?string {
-                                    $author = $record->serviceRequestFormSubmission?->author;
-
-                                    return $author
-                                        ? resolve($author::filamentResource())->getUrl('view', ['record' => $author])
-                                        : null;
-                                })
-                                ->placeholder('-'),
-                            TextEntry::make('serviceRequestFormSubmission.author_type')
-                                ->label('Submitted By Type')
-                                ->badge()
-                                ->color('success')
-                                ->formatStateUsing(fn (?string $state): ?string => filled($state) ? ucfirst($state) : null)
-                                ->placeholder('-'),
-                        ]),
-                    ViewEntry::make('serviceRequestFormSubmission')
-                        ->view('filament.infolists.components.submission-entry'),
-                ]),
-            Section::make('SLA Management')
-                ->visible(fn (ServiceRequest $record): bool => $record->priority?->sla !== null)
-                ->schema([
-                    Group::make([
-                        TextEntry::make('sla_response_seconds')
-                            ->label('Response agreement')
-                            ->state(fn (ServiceRequest $record): ?int => $record->getSlaResponseSeconds())
-                            ->formatStateUsing($formatSecondsAsInterval)
-                            ->placeholder('-'),
-                        TextEntry::make('response_age')
-                            ->label('Response age')
-                            ->state(fn (ServiceRequest $record): int => $record->getLatestResponseSeconds())
-                            ->formatStateUsing($formatSecondsAsInterval)
-                            ->placeholder('-'),
-                        TextEntry::make('response_sla_compliance')
-                            ->label('Response compliance')
-                            ->badge()
-                            ->state(fn (ServiceRequest $record): ?SlaComplianceStatus => $record->getResponseSlaComplianceStatus()),
-                        TextEntry::make('time_to_resolution')
-                            ->label('Time to Resolution')
-                            ->formatStateUsing(function (int $state) {
-                                $interval = Carbon::now()->diffAsCarbonInterval(Carbon::now()->addSeconds($state));
-                                $days = $interval->d;
-                                $hours = $interval->h;
-                                $minutes = $interval->i;
-
-                                return "{$days}d {$hours}h {$minutes}m";
-                            }),
-                    ]),
-                    Group::make([
-                        TextEntry::make('sla_resolution_seconds')
-                            ->label('Resolution agreement')
-                            ->state(fn (ServiceRequest $record): ?int => $record->getSlaResolutionSeconds())
-                            ->formatStateUsing($formatSecondsAsInterval)
-                            ->placeholder('-'),
-                        TextEntry::make('resolution_seconds')
-                            ->label('Resolution age')
-                            ->state(fn (ServiceRequest $record): int => $record->getResolutionSeconds())
-                            ->formatStateUsing($formatSecondsAsInterval)
-                            ->placeholder('-'),
-                        TextEntry::make('resolution_sla_compliance')
-                            ->label('Resolution compliance')
-                            ->badge()
-                            ->state(fn (ServiceRequest $record): ?SlaComplianceStatus => $record->getResolutionSlaComplianceStatus()),
-                    ]),
-                ])
-                ->columns(),
-        ];
-    }
-
-    /**
-     * @return array<Section>
-     */
-    private function filesTabSchema(): array
-    {
-        $uploadsMediaCollection = app(ResolveUploadsMediaCollectionForServiceRequest::class)->__invoke();
-
-        return [
-            Section::make()
-                ->schema(fn (ServiceRequest $record): array => [
-                    Livewire::make(ServiceRequestMediaTable::class, [
-                        'record' => $record,
-                        'collectionName' => $uploadsMediaCollection->getName(),
-                    ]),
-                ]),
-        ];
-    }
 
     /**
      * @return array<Section>
