+++ app-modules/service-management/src/Filament/Resources/ServiceRequests/Schemas/ServiceRequestInfolist.php
@@ -1,0 +1,250 @@
+<?php
+
+/*
+<COPYRIGHT>
+
+    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
+
+    Aiding App® is licensed under the Elastic License 2.0. For more details,
+    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
+
+    Notice:
+
+    - You may not provide the software to third parties as a hosted or managed
+      service, where the service provides users with access to any substantial set of
+      the features or functionality of the software.
+    - You may not move, change, disable, or circumvent the license key functionality
+      in the software, and you may not remove or obscure any functionality in the
+      software that is protected by the license key.
+    - You may not alter, remove, or obscure any licensing, copyright, or other notices
+      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      to applicable law.
+    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
+      same in return. Canyon GBS® and Aiding App® are registered trademarks of
+      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
+      vigorously.
+    - The software solution, including services, infrastructure, and code, is offered as a
+      Software as a Service (SaaS) by Canyon GBS Inc.
+    - Use of this software implies agreement to the license terms and conditions as stated
+      in the Elastic License 2.0.
+
+    For more information or inquiries please visit our website at
+    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
+
+</COPYRIGHT>
+*/
+
+namespace AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Schemas;
+
+use AidingApp\Contact\Filament\Resources\ContactResource;
+use AidingApp\Contact\Models\Contact;
+use AidingApp\Division\Models\Division;
+use AidingApp\ServiceManagement\Actions\ResolveUploadsMediaCollectionForServiceRequest;
+use AidingApp\ServiceManagement\Enums\SlaComplianceStatus;
+use AidingApp\ServiceManagement\Filament\Widgets\ServiceRequestMediaTable;
+use AidingApp\ServiceManagement\Models\ServiceRequest;
+use Carbon\Carbon;
+use Carbon\CarbonInterval;
+use Filament\Infolists\Components\TextEntry;
+use Filament\Infolists\Components\ViewEntry;
+use Filament\Schemas\Components\Grid;
+use Filament\Schemas\Components\Group;
+use Filament\Schemas\Components\Livewire;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Schema;
+use Illuminate\Support\HtmlString;
+
+class ServiceRequestInfolist
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->schema([
+                static::headerSection(),
+                ...static::detailSections(),
+                static::filesSection(),
+                static::feedbackSummarySection(),
+            ]);
+    }
+
+    public static function headerSection(): Section
+    {
+        return Section::make()
+            ->heading(fn (ServiceRequest $record): HtmlString => new HtmlString(
+                view('filament.infolists.components.service-request-heading', [
+                    'serviceRequestNumber' => $record->service_request_number,
+                    'category' => $record->category,
+                    'type' => $record->priority?->type()->first()?->name,
+                ])->render()
+            ))
+            ->schema([
+                TextEntry::make('division.name')
+                    ->visible(fn (): bool => Division::count() > 1)
+                    ->label('Division'),
+                Grid::make(3)
+                    ->schema([
+                        TextEntry::make('respondent')
+                            ->label('Customer Contact')
+                            ->color('primary')
+                            ->html()
+                            ->state(function (ServiceRequest $record): string {
+                                /** @var Contact $respondent */
+                                $respondent = $record->respondent;
+                                $organizationName = $respondent->organization->name ?? 'Unaffiliated';
+
+                                return "{$respondent->{Contact::displayNameKey()}} ({$respondent->type->name})<br>{$organizationName}";
+                            })
+                            ->url(function (ServiceRequest $record) {
+                                /** @var Contact $respondent */
+                                $respondent = $record->respondent;
+
+                                return ContactResource::getUrl('view', ['record' => $respondent->id]);
+                            }),
+                        TextEntry::make('created_at')
+                            ->label('Created')
+                            ->dateTime()
+                            ->hintIcon(null),
+                        TextEntry::make('updated_at')
+                            ->label('Last Updated')
+                            ->dateTime()
+                            ->hintIcon(null),
+                    ])->columns(3),
+                Grid::make(3)
+                    ->schema([
+                        TextEntry::make('status.name')
+                            ->label('Status')
+                            ->badge()
+                            ->color(fn (ServiceRequest $record): string => $record->status->color->value),
+                        TextEntry::make('priority.name')
+                            ->label('Priority'),
+                    ])->columns(3),
+            ])
+            ->columns();
+    }
+
+    /**
+     * @return array<Section>
+     */
+    public static function detailSections(): array
+    {
+        $formatSecondsAsInterval = fn (?int $state): ?string => $state ? CarbonInterval::seconds($state)->cascade()->forHumans(short: true) : null;
+
+        return [
+            Section::make('Title')
+                ->schema([
+                    TextEntry::make('title')
+                        ->hiddenLabel(),
+                ]),
+            Section::make('Description')
+                ->schema([
+                    TextEntry::make('close_details')
+                        ->hiddenLabel()
+                        ->markdown(),
+                ]),
+            Section::make('Form Details')
+                ->visible(fn (ServiceRequest $record): bool => ! is_null($record->serviceRequestFormSubmission))
+                ->schema([
+                    Grid::make(3)
+                        ->schema([
+                            TextEntry::make('serviceRequestFormSubmission.submitted_at')
+                                ->label('Submitted')
+                                ->dateTime(),
+                            TextEntry::make('serviceRequestFormSubmission.author.email')
+                                ->label('Submitted By')
+                                ->color('primary')
+                                ->url(function (ServiceRequest $record): ?string {
+                                    $author = $record->serviceRequestFormSubmission?->author;
+
+                                    return $author
+                                        ? resolve($author::filamentResource())->getUrl('view', ['record' => $author])
+                                        : null;
+                                })
+                                ->placeholder('-'),
+                            TextEntry::make('serviceRequestFormSubmission.author_type')
+                                ->label('Submitted By Type')
+                                ->badge()
+                                ->color('success')
+                                ->formatStateUsing(fn (?string $state): ?string => filled($state) ? ucfirst($state) : null)
+                                ->placeholder('-'),
+                        ]),
+                    ViewEntry::make('serviceRequestFormSubmission')
+                        ->view('filament.infolists.components.submission-entry'),
+                ]),
+            Section::make('SLA Management')
+                ->visible(fn (ServiceRequest $record): bool => $record->priority?->sla !== null)
+                ->schema([
+                    Group::make([
+                        TextEntry::make('sla_response_seconds')
+                            ->label('Response agreement')
+                            ->state(fn (ServiceRequest $record): ?int => $record->getSlaResponseSeconds())
+                            ->formatStateUsing($formatSecondsAsInterval)
+                            ->placeholder('-'),
+                        TextEntry::make('response_age')
+                            ->label('Response age')
+                            ->state(fn (ServiceRequest $record): int => $record->getLatestResponseSeconds())
+                            ->formatStateUsing($formatSecondsAsInterval)
+                            ->placeholder('-'),
+                        TextEntry::make('response_sla_compliance')
+                            ->label('Response compliance')
+                            ->badge()
+                            ->state(fn (ServiceRequest $record): ?SlaComplianceStatus => $record->getResponseSlaComplianceStatus()),
+                        TextEntry::make('time_to_resolution')
+                            ->label('Time to Resolution')
+                            ->formatStateUsing(function (int $state) {
+                                $interval = Carbon::now()->diffAsCarbonInterval(Carbon::now()->addSeconds($state));
+                                $days = $interval->d;
+                                $hours = $interval->h;
+                                $minutes = $interval->i;
+
+                                return "{$days}d {$hours}h {$minutes}m";
+                            }),
+                    ]),
+                    Group::make([
+                        TextEntry::make('sla_resolution_seconds')
+                            ->label('Resolution agreement')
+                            ->state(fn (ServiceRequest $record): ?int => $record->getSlaResolutionSeconds())
+                            ->formatStateUsing($formatSecondsAsInterval)
+                            ->placeholder('-'),
+                        TextEntry::make('resolution_seconds')
+                            ->label('Resolution age')
+                            ->state(fn (ServiceRequest $record): int => $record->getResolutionSeconds())
+                            ->formatStateUsing($formatSecondsAsInterval)
+                            ->placeholder('-'),
+                        TextEntry::make('resolution_sla_compliance')
+                            ->label('Resolution compliance')
+                            ->badge()
+                            ->state(fn (ServiceRequest $record): ?SlaComplianceStatus => $record->getResolutionSlaComplianceStatus()),
+                    ]),
+                ])
+                ->columns(),
+        ];
+    }
+
+    public static function filesSection(): Section
+    {
+        $uploadsMediaCollection = app(ResolveUploadsMediaCollectionForServiceRequest::class)->__invoke();
+
+        return Section::make('Files')
+            ->schema(fn (ServiceRequest $record): array => [
+                Livewire::make(ServiceRequestMediaTable::class, [
+                    'record' => $record,
+                    'collectionName' => $uploadsMediaCollection->getName(),
+                ]),
+            ]);
+    }
+
+    public static function feedbackSummarySection(): Section
+    {
+        return Section::make('Feedback')
+            ->visible(fn (ServiceRequest $record): bool => $record->feedback()->exists())
+            ->schema([
+                TextEntry::make('feedback.csat_answer')
+                    ->label('CSAT')
+                    ->badge(),
+                TextEntry::make('feedback.nps_answer')
+                    ->label('NPS')
+                    ->badge(),
+            ])
+            ->columns();
+    }
+}
